soap bark + my own edits
🟥 Soap Bark 🟥 (Zach), [2/12/21 2:44 PM]
"The challenge is to ensure this mechanism maintains a cost-of-attack hight than the possible rewards"

Is this supposed to be, "higher than"?

🟥 Soap Bark 🟥 (Zach), [2/12/21 2:50 PM]
" giving storage price on the chain this is a market determined price rather than an incidental one."   "This" should be omitted.

🟥 Soap Bark 🟥 (Zach), [2/12/21 2:52 PM]
"Should they find a solution it is propagated into the network as a normal fee-paying transaction."

^This needs a comma after, "solution"

🟥 Soap Bark 🟥 (Zach), [2/12/21 2:53 PM]
"Only one solution may be included in any block, and the solution must be to the previous block for the current block to be considered valid."

"must be directed towards"  - maybe that sounds better

🟥 Soap Bark 🟥 (Zach), [2/12/21 2:55 PM]
"they would have to spend proportionally *more* than their increased odds are worth, making attacks based on receiving one's own transactions a losing strategy."

"ensuring" vs "making" sounds a bit better imo.

🟥 Soap Bark 🟥 (Zach), [2/12/21 2:58 PM]
"Let us call this method "powsplit". "

"This method can be referred to as, "powsplit".

🟥 Soap Bark 🟥 (Zach), [2/12/21 3:00 PM]
"The remainder that would have been returned to miners should a golden ticket solution have been found"

"Assuming a golden ticket solution has been found, the remainder that would have been returned to miners gets distributed......"

🟥 Soap Bark 🟥 (Zach), [2/12/21 3:01 PM]
"in a similar method to the selection of"

"in a similar fashion as the selection of"

🟥 Soap Bark 🟥 (Zach), [2/12/21 3:02 PM]
"To avoid throttling attacks on the staking mechanism it is wise to disallow stakers from spending or withdrawing staked UTXOs until they have received payment."

Need a comma after, "mechanism"

🟥 Soap Bark 🟥 (Zach), [2/12/21 3:04 PM]
"The consensus variables of the network are updated when and only when golden tickets are solved and included in the blockchain."

"The consensus variables of the network are only updated when golden tickets are solved and included in the blockchain."

(redundancy)

🟥 Soap Bark 🟥 (Zach), [2/12/21 3:06 PM]
"Users making votes are also withholding their fees from nodes voting differently to themselves."

"casting" instead of "making"

🟥 Soap Bark 🟥 (Zach), [2/12/21 3:07 PM]
Ok that is it.  Some changes are more important than others.  Thank you for giving me an opportunity to help out!

I did some other small changes, and a couple larger ones like clarifying reward distribution at the start of the golden ticket section and adding some clarification to the intro powsplit paragraph.